+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0]
+### Added
+- New support for querying variables (UMM-V)
+### Changed
+- Can now import `ToolQuery` `ServiceQuery` `VariableQuery` straight from cmr module. (e.g. `from cmr import ToolQuery`)
+
 ## [0.5.0]
 ### Added
 - New support for querying tools (UMM-T) and services (UMM-S)
@@ -16,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Older]
 - Prior releases of this software originated from https://github.com/jddeal/python-cmr/releases
 
-[Unreleased]: https://github.com/nasa/python_cmr/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/nasa/python_cmr/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/nasa/python_cmr/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/nasa/python_cmr/compare/ef0f9e7d67ce99d342a568bd6a098c3462df16d2...v0.5.0
 [Older]: https://github.com/jddeal/python-cmr/releases